Restructure restaurants into selection
