--- src/Components/Restaurants/Restaurant.css
@@ -1,6 +1,0 @@
-.restaurants-container {
-    padding: 20px;
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    grid-gap: 20px;
-  }
