+++ src/Components/Selection/Selection.css
@@ -1,1 +1,6 @@
-
+.restaurants-container {
+  padding: 20px;
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-gap: 20px;
+}
